requirements.txt
streamlit
pandas
python-dotenv
google-generativeai
openpyxl
pandas
google-generativeai
